Separate lock.h: Linux
